refactor: change names of files and functions to follow naming conventions
