+++ .gitignore
@@ -1,3 +1,3 @@
-a.out
+**/*.out
 build/
